+++ esbuild.config.mjs
@@ -116,6 +116,7 @@ async function buildBackground() {
     readFile("modules/backup-action-handler.js"),
     readFile("modules/organization-action-handler.js"),
     readFile("modules/settings-action-handler.js"),
+    readFile("modules/security-action-handler.js"),
     readFile("modules/message-router.js"),
     readFile("modules/gm-audio-handler.js"),
     readFile("modules/gm-menu-handler.js"),